+++ lib/src/filter.dart
@@ -3,80 +3,119 @@
 
 part of sortedmap;
 
-/// The signature of a generic predicate function.
-///
-/// A [Predicate] function returns true if [v] matches the predicate.
-///
-typedef bool Predicate<T>(T v);
 
 /// Defines a filter that can be used with [FilteredMap] objects.
 ///
 /// A [Filter] defines:
 ///
-/// * an ordering through the [compare] method
-/// * a predicate through the [isValid] method
+/// * an [ordering]
 /// * a [limit] of the number of elements
+/// * a [validInterval], i.e. an upper and lower limit (inclusive) of valid values.
 ///
-/// The [reverse] property determines if the limit is applied from the start or
+/// The [reversed] property determines if the limit is applied from the start or
 /// from the end.
 ///
-class Filter<T> {
+class Filter<K extends Comparable,V> {
 
-  final Comparator<T> _compare;
-  final Predicate<T> _isValid;
+  /// The ordering.
+  final Ordering<K,V> ordering;
 
   /// The number of elements allowed in a set.
   final int limit;
 
   /// Determines if the limit is applied from the start or from the end.
-  final bool reverse;
+  final bool reversed;
+
+  /// The interval of acceptable values.
+  final KeyValueInterval<K,Comparable> validInterval;
 
   /// Creates a new [Filter] instance.
-  const Filter({Comparator<T> compare, Predicate<T> isValid, this.limit,
-  this.reverse: false}) : _compare = compare, _isValid = isValid;
-
-  /// Compares two items.
-  int compare(T a, T b) => _compare==null ? Comparable.compare(a as Comparable,b as Comparable) : _compare(a,b);
-
-  /// Checks if an item is valid.
-  bool isValid(T v) => (_isValid ?? (_)=>true)(v);
+  const Filter({this.ordering: const Ordering.byKey(),
+    this.limit, this.reversed: false, this.validInterval: const KeyValueInterval()});
 
   @override
-  int get hashCode => quiver.hash4(compare, isValid, limit, reverse);
+  int get hashCode => quiver.hash4(ordering, limit, reversed, validInterval);
 
   @override
-  bool operator==(dynamic other) => other is Filter&&other.compare==compare
-      &&other.isValid==isValid&&other.limit==limit&&other.reverse==reverse;
+  bool operator==(dynamic other) => other is Filter&&other.ordering==ordering
+      &&other.limit==limit&&other.reversed==reversed&&other.validInterval==validInterval;
 }
 
-/// A predicate function that accepts all values between (including) a lower and
-/// upper limit.
-class Range<T> extends Function {
+/// Defines an interval.
+class KeyValueInterval<K extends Comparable,V extends Comparable> {
+
+  final K _startKey;
+  final V _startValue;
+  final K _endKey;
+  final V _endValue;
+
+  /// Creates a new interval from individual keys and values.
+  const KeyValueInterval([this._startKey, this._startValue, this._endKey, this._endValue]);
+
+  /// Creates a new interval from key/value pairs.
+  factory KeyValueInterval.fromPairs(Pair<K,V> start, Pair<K,V> end) =>
+      new KeyValueInterval(start.key, start.value, end.key, end.value);
 
   /// The lower limit.
-  final T start;
+  Pair<K,V> get start => new Pair.min(_startKey,_startValue);
 
   /// The upper limit.
-  final T end;
+  Pair<K,V> get end => new Pair.max(_endKey,_endValue);
 
-  final Comparator<T> _comparator;
+  /// Returns true if this interval is unbounded.
+  bool get isUnlimited => start.key==null&&start.value==null&&end.key==null&&end.value==null;
 
-  /// Creates a new range wit lower limit [start] and upper limit [end].
-  /// An optional [comparator] function defines the ordering.
-  Range(this.start, this.end, [this._comparator]);
+  /// Creates a new interval replacing the lower limit.
+  KeyValueInterval<K,V> startAt(K key, V value) =>
+      new KeyValueInterval(key, value, _endKey, _endValue);
 
-  /// The [Comparator] function that defines the ordering.
-  Comparator<T> get comparator => (_comparator ?? Comparable.compare as Comparator<T>);
+  /// Creates a new interval replacing the upper limit.
+  KeyValueInterval<K,V> endAt(K key, V value) =>
+      new KeyValueInterval(_startKey, _startValue, key, value);
 
-  /// Checks if a [value] is between the lower and upper limit.
-  bool call(T value) => (start==null || comparator(start, value) <= 0)
-      && (end==null || comparator(value, end) <=0);
+
+  /// Returns true if `point` is within the bounds (inclusive) of this interval.
+  bool containsPoint(Pair<K,V> p) => _isAfterStart(p.key,p.value)&&_isBeforeEnd(p.key,p.value);
+
+  bool _isAfterStart(K key, V value) {
+    if (start.value==null) {
+      if (value!=null) return true;
+    } else {
+      if (value==null) return false;
+      var cmp = Comparable.compare(start.value,value);
+      if (cmp>0) return false;
+      if (cmp<0) return true;
+    }
+    if (start.key==null) return true;
+    if (key==null) return false;
+    return (Comparable.compare(start.key,key))<=0;
+  }
+
+  bool _isBeforeEnd(K key, V value) {
+    if (end.value==null) {
+      if (value!=null) return true;
+    } else {
+      if (value==null) return false;
+      var cmp = Comparable.compare(end.value,value);
+      if (cmp>0) return true;
+      if (cmp<0) return false;
+    }
+    if (end.key==null) return true;
+    if (key==null) return false;
+    return (Comparable.compare(end.key,key))>=0;
+  }
+
+  /// Returns true if this interval contains the interval `other`.
+  bool contains(KeyValueInterval<K,V> other) =>
+      containsPoint(other.start)&&containsPoint(other.end);
 
   @override
-  int get hashCode => quiver.hash3(start,end,comparator);
+  int get hashCode => quiver.hash2(start,end);
 
   @override
-  bool operator==(dynamic other) => other is Range&&other.start==start
-      &&other.end==end&&other.comparator==comparator;
+  bool operator==(dynamic other) => other is KeyValueInterval&&other.start==start&&
+  other.end==end;
+
+  @override
+  String toString() => "KeyValueInterval[$start,$end]";
 }
-