+++ lib/src/sortedmap.dart
@@ -4,57 +4,18 @@
 part of sortedmap;
 
 
-/// A [Pair] represents a key/value pair.
-class Pair<K,V> implements Comparable<Pair<K,V>> {
-
-  /// The key.
-  final K key;
-
-  /// The value.
-  final V value;
-
-  /// Creates a new key/value pair.
-  const Pair(this.key,this.value);
-
-  @override
-  int get hashCode => quiver.hash2(key,value);
-
-  @override
-  bool operator==(dynamic other) => other is Pair && other.key==key && other.value == value;
-
-  @override
-  int compareTo(Pair<K, V> other) => Comparable.compare(key as Comparable,
-      other.key as Comparable);
-}
 
 /// A [Map] of objects that can be ordered relative to each other.
-///
-/// Key/value pairs of the map are compared using the `compare` function
-/// passed in the constructor.
-///
-/// If the compare function is omitted, the objects are ordered by key which
-/// is assumed to be [Comparable], and are compared using their
-/// [Comparable.compareTo] method.
-/// Non-comparable objects (including `null`) will not work as keys
-/// in that case.
-class SortedMap<K,V> extends MapBase<K,V> /*implements Differentiable<SortedMap<K,V>>*/ {
+abstract class SortedMap<K extends Comparable,V> implements Map<K,V> {
 
-  /// A [Comparator] function that defines the ordering.
-  final Comparator<Pair<K,V>> comparator;
+  /// Creates a new [SortedMap] instance with an optional ordering.
+  factory SortedMap([Ordering<K,V> ordering = const Ordering.byKey()]) =>
+      new _SortedMap._(ordering, null, null);
 
-  SplayTreeSet<Pair<K,V>> _sortedPairs;
-  Map<K,V> _map = {};
-
-  /// Creates a new [SortedMap] instance with an optional [compare] function,
-  /// defining the ordering.
-  SortedMap([int compare(Pair<K,V> a, Pair<K,V> b)]) :
-        _sortedPairs = new SplayTreeSet(compare), comparator = compare;
-
-  SortedMap._(this.comparator, this._sortedPairs, this._map);
 
   /// Creates a [SortedMap] that contains all key/value pairs of [other].
-  factory SortedMap.from(Map<K,V> other, [int compare(Pair<K,V> a, Pair<K,V> b)]) {
-    return new SortedMap(compare)..addAll(other);
+  factory SortedMap.from(Map<K,V> other, [Ordering<K,V> ordering = const Ordering.byKey()]) {
+    return new SortedMap(ordering)..addAll(other);
   }
 
   /// Creates a [SortedMap] where the keys and values are computed from the
@@ -70,9 +31,9 @@ class SortedMap<K,V> extends MapBase<K,V> /*implements Differentiable<SortedMap<
   /// use the iterable value itself.
   factory SortedMap.fromIterable(Iterable iterable,
       {K key(element),
-      V value(element),
-      int compare(Pair<K,V> a, Pair<K,V> b)}) {
-    SortedMap<K, V> map = new SortedMap<K, V>(compare);
+        V value(element),
+        Ordering<K,V> ordering: const Ordering.byKey()}) {
+    SortedMap<K, V> map = new SortedMap<K, V>(ordering);
 
     if (key == null) key = (K v)=>v;
     if (value == null) value = (V v)=>v;
@@ -93,8 +54,8 @@ class SortedMap<K,V> extends MapBase<K,V> /*implements Differentiable<SortedMap<
   ///
   /// It is an error if the two [Iterable]s don't have the same length.
   factory SortedMap.fromIterables(Iterable<K> keys, Iterable<V> values,
-      [int compare(Pair<K,V> a, Pair<K,V> b)]) {
-    SortedMap<K, V> map = new SortedMap<K, V>(compare);
+      [Ordering<K,V> ordering = const Ordering.byKey()]) {
+    SortedMap<K, V> map = new SortedMap<K, V>(ordering);
     Iterator<K> keyIterator = keys.iterator;
     Iterator<V> valueIterator = values.iterator;
 
@@ -113,26 +74,72 @@ class SortedMap<K,V> extends MapBase<K,V> /*implements Differentiable<SortedMap<
     return map;
   }
 
-  /// Makes a copy of this map. The key/value pairs in the map are not cloned.
-  SortedMap<K,V> clone() => new SortedMap._(this.comparator, _sortedPairs.toSet(), new Map.from(_map));
+  /// The ordering.
+  Ordering<K,V> get ordering;
 
-  /// Gets the key/value pairs as an iterable
-  Iterable<Pair<K,V>> get pairs => _sortedPairs.toSet();
+  /// Makes a copy of this map. The key/value pairs in the map are not cloned.
+  SortedMap<K,V> clone();
+
+  /// Get the last key in the map for which the key/value pair is strictly
+  /// smaller than that of [key]. Returns [:null:] if no key was not found.
+  K lastKeyBefore(K key);
+
+  /// Get the first key in the map for which the key/value pair is strictly
+  /// larger than that of [key]. Returns [:null:] if no key was not found.
+  K firstKeyAfter(K key);
+
+  /// Gets the keys within the desired bounds and limit.
+  Iterable<K> subkeys({Pair<K,Comparable> start, Pair<K,Comparable> end, int limit, bool reversed: false});
+
+  /// Creates a filtered view of this map.
+  FilteredMapView<K,V> filteredMapView({Pair<K,Comparable> start, Pair<K,Comparable> end, int limit, bool reversed: false}) =>
+      new FilteredMapView(this, start: start, end: end, limit: limit, reversed: reversed);
+
+  /// Creates a new filtered map based on this map.
+  FilteredMap<K,V> filteredMap({Pair<K,Comparable> start, Pair<K,Comparable> end, int limit, bool reversed: false}) =>
+      new FilteredMap(new Filter(validInterval: new KeyValueInterval.fromPairs(start, end), ordering: ordering, limit: limit, reversed: reversed))..addAll(this);
+
+
+  Pair<K,Comparable> _pairForKey(K key) => containsKey(key) ? ordering.mapKeyValue(key, this[key]) : null;
+
+}
+
+
+class _SortedMap<K extends Comparable,V> extends MapBase<K,V> with SortedMap<K,V> {
+
+  @override
+  final Ordering<K,V> ordering;
+
+  TreeSet<Pair<K,Comparable>> _sortedPairs;
+  Map<K,V> _map;
+
+  _SortedMap._(this.ordering, this._sortedPairs, this._map) {
+    _sortedPairs ??= new TreeSet();
+    _map ??= {};
+  }
+
+  @override
+  bool containsKey(Object key) => _map.containsKey(key);
+
+  @override
+  Iterable<K> get keys => _sortedPairs.map((p)=>p.key);
+
+  @override
+  SortedMap<K,V> clone() => new _SortedMap<K,V>._(this.ordering, new TreeSet()..addAll(_sortedPairs), new Map<K,V>.from(_map));
 
   @override
   V operator [](Object key) => _map[key];
 
+
   @override
   void operator []=(K key, V value) {
-    if (_map.containsKey(key)) {
-      _sortedPairs.remove(new Pair(key, _map[key]));
-    }
-    _addPair(new Pair(key,value));
+    _sortedPairs.remove(_pairForKey(key));
+    _addPair(key,value);
   }
 
-  void _addPair(Pair<K,V> p) {
-    _map[p.key] = p.value;
-    _sortedPairs.add(p);
+  void _addPair(K key, V value) {
+    _map[key] = value;
+    _sortedPairs.add(ordering.mapKeyValue(key,value));
   }
 
   @override
@@ -141,39 +148,69 @@ class SortedMap<K,V> extends MapBase<K,V> /*implements Differentiable<SortedMap<
     _sortedPairs.clear();
   }
 
-  @override
-  Iterable<K> get keys => pairs.map((p)=>p.key);
 
   @override
   V remove(Object key) {
-    if (_map.containsKey(key)) {
-      _sortedPairs.remove(new Pair(key, _map[key]));
-    }
+    if (!_map.containsKey(key)) return null;
+    _sortedPairs.remove(_pairForKey(key as K));
     return _map.remove(key);
   }
 
-  /// Get the last key in the map for which the key/value pair is strictly
-  /// smaller than that of [key]. Returns [:null:] if no key was not found.
+  @override
   K lastKeyBefore(K key) {
     if (!_map.containsKey(key)) {
       throw new StateError("No such key $key in collection");
     }
-    var pair = new Pair(key, _map[key]);
-    return pairs.takeWhile((p)=>p!=pair)
-        .map((p)=>p.key)
-        .lastWhere((_)=>true, orElse: ()=>null);
+    var pair = _pairForKey(key);
+    var it = _sortedPairs.fromIterator(pair, reversed: true);
+    while (it.moveNext()&&it.current==pair);
+    return it.current?.key;
   }
 
-  /// Get the first key in the map for which the key/value pair is strictly
-  /// larger than that of [key]. Returns [:null:] if no key was not found.
+  @override
   K firstKeyAfter(K key) {
     if (!_map.containsKey(key)) {
       throw new StateError("No such key $key in collection");
     }
-    var pair = new Pair(key, _map[key]);
-    return pairs.skipWhile((p)=>p!=pair).skip(1)
-        .map((p)=>p.key)
-        .firstWhere((_)=>true, orElse: ()=>null);
+    var pair = _pairForKey(key);
+    var it = _sortedPairs.fromIterator(pair);
+    while (it.moveNext()&&it.current==pair);
+    return it.current?.key;
   }
+
+  @override
+  Iterable<K> subkeys({Pair<K,Comparable> start, Pair<K,Comparable> end, int limit, bool reversed: false}) {
+    var it = _subkeys(start, end, limit, reversed);
+    if (reversed) return it.toList().reversed;
+    return it;
+  }
+
+  Iterable<K> _subkeys(Pair<K,Comparable> start, Pair<K,Comparable> end, int limit, bool reversed) sync* {
+    var from = reversed ? end : start;
+    Iterator it = _sortedPairs.fromIterator(from, reversed: reversed);
+    int count = 0;
+    while (it.moveNext()&&(limit==null||count++<limit)) {
+      var cmp = Comparable.compare(it.current, reversed ? start : end);
+      if ((reversed&&cmp<0)||(!reversed&&cmp>0)) return;
+      yield it.current.key;
+    }
+  }
+
+}
+
+/// An unmodifiable sorted map.
+abstract class UnmodifiableSortedMap<K extends Comparable,V> implements SortedMap<K,V> {
+  @override
+  void operator []=(K key, V value) =>
+      throw new UnsupportedError("Map view cannot be modified.");
+
+  @override
+  void clear() =>
+      throw new UnsupportedError("Map view cannot be modified.");
+
+  @override
+  V remove(Object key) =>
+      throw new UnsupportedError("Map view cannot be modified.");
+
 
 }