+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.2.0
+
+- new FilteredMapView class
+- some api changes
+
+## 0.1.2
+
+- keep elements after limit in FilteredMap
+
 ## 0.1.1
 
 - `compare` and `isValid` methods iso getters in `Filter`